8.0.0/jasperserver/jasperserver-dto/src/main/java/com/jaspersoft/jasperserver/dto/adhoc/query/el/adapters/TimeToStringXmlAdapter.java
/*
 * Copyright (C) 2005 - 2020 TIBCO Software Inc. All rights reserved.
 * http://www.jaspersoft.com.
 *
 * Unless you have purchased a commercial license agreement from Jaspersoft,
 * the following license terms apply:
 *
 * This program is free software: you can redistribute it and/or modify
 * it under the terms of the GNU Affero General Public License as
 * published by the Free Software Foundation, either version 3 of the
 * License, or (at your option) any later version.
 *
 * This program is distributed in the hope that it will be useful,
 * but WITHOUT ANY WARRANTY; without even the implied warranty of
 * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
 * GNU Affero General Public License for more details.
 *
 * You should have received a copy of the GNU Affero General Public License
 * along with this program. If not, see <http://www.gnu.org/licenses/>.
 */
package com.jaspersoft.jasperserver.dto.adhoc.query.el.adapters;

import javax.xml.bind.annotation.adapters.XmlAdapter;
import java.sql.Time;
import java.text.SimpleDateFormat;

/**
 * @author Grant Bacon <gbacon@tibco.com>
 * @author Stas Chubar <schubar@tibco.com>
 * @version $Id $
 */
public class TimeToStringXmlAdapter extends XmlAdapter<String, Time> {

    private static final SimpleDateFormat DATE_FORMAT = DomELCommonSimpleDateFormats.timeFormat();

    @Override
    public String marshal(Time v) throws Exception {
        return DATE_FORMAT.format(v);
    }

    @Override
    public Time unmarshal(String v) throws Exception {
        return new Time(DATE_FORMAT.parse(v).getTime());
    }
}
